mediacapture-streams/MediaStream-supported-by-feature-policy.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('camera', document.featurePolicy.features());
}, 'document.featurePolicy.features should advertise camera.');

test(() => {
    assert_in_array('microphone', document.featurePolicy.features());
}, 'document.featurePolicy.features should advertise microphone.');
</script>
